Split embedded integration tests in CI to separate job
